Move coursier.util.Repositories to coursier module
and under coursier namespace
