+++ modules/cli/src/main/scala-2.12/coursier/cli/params/shared/RepositoryParams.scala
@@ -1,12 +1,12 @@
 package coursier.cli.params.shared
 
 import cats.data.{NonEmptyList, Validated, ValidatedNel}
+import coursier.Repositories
 import coursier.cli.options.shared.RepositoryOptions
 import coursier.core.Repository
 import coursier.ivy.IvyRepository
 import coursier.maven.MavenRepository
 import coursier.parse.RepositoryParser
-import coursier.util.Repositories
 
 
 object RepositoryParams {
