+++ modules/coursier/js/src/main/scala/coursier/PlatformResolve.scala
@@ -1,7 +1,5 @@
 package coursier
 
-import coursier.util.Repositories
-
 abstract class PlatformResolve {
 
   val defaultRepositories: Seq[Repository] =
