+++ modules/coursier/shared/src/main/scala/coursier/Repositories.scala
@@ -1,7 +1,6 @@
-package coursier.util
+package coursier
 
 import coursier.ivy.IvyRepository
-import coursier.maven.MavenRepository
 
 object Repositories {
   def central: MavenRepository =
